+++ core/metrics/config.go
@@ -15,7 +15,7 @@ package metrics
 import (
 	"time"
 
-	"code.vegaprotocol.io/vega/core/config/encoding"
+	"code.vegaprotocol.io/vega/libs/config/encoding"
 	"code.vegaprotocol.io/vega/logging"
 )