+++ libs/http/rate_limit.go
@@ -19,7 +19,7 @@ import (
 	"sync"
 	"time"
 
-	"code.vegaprotocol.io/vega/core/config/encoding"
+	"code.vegaprotocol.io/vega/libs/config/encoding"
 )
 
 type RateLimitConfig struct {